backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
dj-database-url==2.1.0
